+++ routes/index.js
@@ -2,17 +2,15 @@ const router = require('express').Router();
 
 const userRouter = require('./users');
 const articleRouter = require('./articles');
-const authentication = require('../middlewares/authentication');
-const authorization = require('../middlewares/authorization');
+const auth = require('../middlewares/auth');
 const { login, createUser } = require('../controllers/users');
 const {
-  validateSignInBody, validateSignUpBody, validateAuthHeaders,
+  validateSignInBody, validateSignUpBody, validateAuthCookies,
 } = require('../middlewares/validations');
 
 router.post('/signin', validateSignInBody, login);
 router.post('/signup', validateSignUpBody, createUser);
-router.use(validateAuthHeaders, authentication);
-router.use(authorization);
+router.use(validateAuthCookies, auth);
 router.use('/users', userRouter);
 router.use('/articles', articleRouter);
